Rename compat to record
Going to make a dataclass module too - no other changes other than
renaming this
